Add files via upload
Welcome to my repository! Here, I've created a C program that solves various mathematical problems related to Arithmetic Progressions (AP) from the NCERT Class 10 Math curriculum. This program serves as a handy tool for students studying this topic, providing solutions to questions involving finding terms, common differences, and sums of terms in arithmetic progressions.

This program helps students understand the concepts of Arithmetic Progressions and provides them with a quick and reliable way to solve related questions. It's a great learning resource for anyone studying NCERT Class 10 Math and wanting to grasp the fundamentals of AP effectively.

Feel free to explore the code, use it to solve problems, and contribute to its enhancement. Your contributions are highly appreciated!"